Do not move orverride windows (fixes bug 3753)
This is a temporary fix. Proper solution for number of issues that pertain to
managing OR windows is being worked on upstream (for details see the upstream
bugzilla at http://bugzilla.gnome.org/show_bug.cgi?id=582639) but this requires
merging in  additional upstream changes first and extensive testing due to its
pervasive nature. I will do so, as soon as possible, but for now this simple
fix addresses bug 3753.