[MutterWindow] added "window-destroyed" signal
This signal gets emitted upon entry into mutter_window_destroyed() before the
MutterWindow object is modified. It can be used by a plugin to track when a
window is removed. Unlike the plugin destroy() vfunction, this signal gets
emitted regardless whether the window is currently visible to the compositor
or not, and unlike the ClutterActor::destroy signal, this signal is emitted
early, so that all the MutterWindow and MetaWindow resources are still valid.