[MutterShadow] An API allowing plugins to generate their custom shadows
The commmit adds new vfunction get_shadow() to MutterPluginClass, allowing
plugins to replace the stock gaussian shadow with a custom actor. The
shadow is encapsulated in a MutterShadow structure that contains the custom
actor, and four points of attachment describing the position and size of the
shadow relative to the window.

In addition, this commit replaces the MutterWindow::no-frame boolean property
with shape-type emum property which allows to force and disable the shadow
on a given window, or leave the compositor to decide.

The plugin API version number is bumped by one.