Infrastructure allowing plugins to implement custom window constraints
This commit adds a constrain_window() vfunction to the plugin class, and hooks
it onto the WM constraint mechanism.

The constraints.h file is split into a public and private parts; the definition
of the theme-specific version of MetaFrameGeometry struct was moved from its
location in theme.h to frames.h to avoid redefinition when autogenerating type
info for enums.

Plugin API version is bumped to 4