[MutterWindow] Remove invalid asserts
When a window is moved from one workspace onto another, its visibility status
can change while it is enqueued for visibility calculation. For example, window
can be initially visible, then enqued to be hidden due to change of its
workspace, and then enqueued again (before the original queue request has been
processed) to be shown due to the change of active workspace. The net effect of
this is that the mutter_window_show() and mutter_window_hide() can be called
on the window that is already in the correct state.

I can't think of a way to fix this without introducing further complexity to
the queue (i.e., so that two subsequent queue requests can cancel out). This
commit changes the two functions to simply return immediately if this scenario
happens.