make sure we create a pixmap for all new mutter windows
In the case where a mutter window is created for an X Window that is
already mapped then we didn't call mutter_window_mark_for_repair and
so we didn't call XCreateWindowPixmap e.g. for menu windows.

Previously this didn't matter because as soon as any damage came through
for the window we'd notice and create the pixmap but damage doesn't
result in a call to check_needs_repair anymore.

This patch renames check_needs_repair and mutter_window_mark_for_repair
to check_needs_pixmap and mutter_window_queue_create_pixmap respectively
to better reflect what they now do. We now also call
mutter_window_queue_create_pixmap in mutter_window_new when the given
Window is already mapped.