Only bail out on managing a specific window for valid reasons, MB#21
MB#21 was due to meta_window_new() failing as a result of catching BadColor
error when querying the window attributes; in fact this error was left over
from a previous operations on a different window. This commit makes creation of
MetaWindow bit more robust so we bail out only on errors that are valid for the
XGetWindowAttributes() call.