Fix frame flags for windows that are created in fullscreen state
Windows created in fullscreen state have the fullscreen_after_placement flag
set until they are constrained, at which point this is translated into the
regular fullscreen flag. Since the frame geometry is fed as an input into the
constrainst function, we need to ensure that the frame META_FRAME_FULLSCREEN
flag is set also for windows that are marked for fullscreening after placement.